Use non-DOMjudge provided FPM config for DOMjudge 5.x as we don't have it yet there. Fixes #35

+++ docker/domserver/scripts/start.sh
@@ -79,19 +79,25 @@ echo "[ok] Changed restapi URL from http://localhost/domjudge to http://localhos
 echo "[..] Copying webserver config"
 # Set up vhost
 cp etc/nginx-conf /etc/nginx/sites-enabled/default
-# Replace nginx php socket location
-sed -i 's/server unix:.*/server unix:\/var\/run\/php-fpm-domjudge.sock;/' /etc/nginx/sites-enabled/default
-# Remove default FPM pool config and link in DOMJudge version
-if [[ -f /etc/php/7.0/fpm/pool.d/www.conf ]]
+if [[ -f /opt/domjudge/domserver/etc/domjudge-fpm.conf ]]
 then
-	rm /etc/php/7.0/fpm/pool.d/www.conf
+	# Replace nginx php socket location
+	sed -i 's!server unix:.*!server unix:/var/run/php-fpm-domjudge.sock;!' /etc/nginx/sites-enabled/default
+	# Remove default FPM pool config and link in DOMJudge version
+	if [[ -f /etc/php/7.0/fpm/pool.d/www.conf ]]
+	then
+		rm /etc/php/7.0/fpm/pool.d/www.conf
+	fi
+	if [[ ! -f /etc/php/7.0/fpm/pool.d/domjudge.conf ]]
+	then
+		ln -s /opt/domjudge/domserver/etc/domjudge-fpm.conf /etc/php/7.0/fpm/pool.d/domjudge.conf
+	fi
+	# Change pm.max_children
+	sed --follow-symlinks -i "s/^pm\.max_children = .*$/pm.max_children = ${FPM_MAX_CHILDREN}/" /etc/php/7.0/fpm/pool.d/domjudge.conf
+else
+	# Replace nginx php socket location
+	sed -i 's!server unix:.*!server unix:/var/run/php/php7.0-fpm.sock;!' /etc/nginx/sites-enabled/default
 fi
-if [[ ! -f /etc/php/7.0/fpm/pool.d/domjudge.conf ]]
-then
-	ln -s /opt/domjudge/domserver/etc/domjudge-fpm.conf /etc/php/7.0/fpm/pool.d/domjudge.conf
-fi
-# Change pm.max_children
-sed -i "s/^pm\.max_children = .*$/pm.max_children = ${FPM_MAX_CHILDREN}/" /etc/php/7.0/fpm/pool.d/domjudge.conf
 
 # Set up permissions
 chown www-data: etc/dbpasswords.secret
